NHL Game Content

Returns game content for a given game ID

nhl_game_content(game_id)

Arguments

game_id
Game unique ID

Value

Returns a tibble

Examples

# \donttest{
  nhl_game_content(game_id=2021020182)
#>    type      id                     date                           title
#> 1 video 9467530 2021-11-09T19:00:00-0500 Marchand uses body to score PPG
#> 2 video 9468158 2021-11-09T19:00:00-0500        Bergeron volleys in goal
#> 3 video 9467963 2021-11-09T19:00:00-0500        Zaitsev scores from wall
#> 4 video 9467774 2021-11-09T19:00:00-0500  Forbort snipes in leading goal
#> 5 video 9466643 2021-11-09T19:00:00-0500          Sanford taps in opener
#> 6 video 9468086 2021-11-09T19:00:00-0500    Murray makes epic stick save
#>                                         blurb
#> 1     OTT@BOS: Marchand scores fortuitous PPG
#> 2     OTT@BOS: Bergeron whips in leading goal
#> 3 OTT@BOS: Zaitsev ties game with sneaky shot
#> 4      OTT@BOS: Forbort takes lead with snipe
#> 5 OTT@BOS: Sanford opens scoring for Senators
#> 6  OTT@BOS: Murray denies Pastrnak with stick
#>                                                                                                                  description
#> 1        David Pastrnak collects the puck and shoots, then it deflects off the chest of Brad Marchand and into net for a PPG
#> 2 David Pastrnak fires a pass to Patrice Bergeron, who settles the puck with his skate then volleys in a shot to make it 3-2
#> 3    Nikita Zaitsev corrals the puck at the half-wall and fires a shot that just gets by Jeremy Swayman to tie the game, 2-2
#> 4                   A nice team play by the Bruins is capped off by Derek Forbort's brilliant wrist shot goal to make it 2-1
#> 5            The Senators stay hard on the forecheck as Brady Tkachuk sends a pass to Zach Sanford to score the opening goal
#> 6                                          Matt Murray makes a desperation save with his stick to deny David Pastrnak a goal
#>   duration authFlow mediaPlaybackId    mediaState
#> 1    00:39    FALSE         9467530 MEDIA_ARCHIVE
#> 2    00:44    FALSE         9468158 MEDIA_ARCHIVE
#> 3    00:28    FALSE         9467963 MEDIA_ARCHIVE
#> 4    00:47    FALSE         9467774 MEDIA_ARCHIVE
#> 5    00:42    FALSE         9466643 MEDIA_ARCHIVE
#> 6    00:30    FALSE         9468086 MEDIA_ARCHIVE
#>                                                                                                                                                                         keywords
#> 1 shareable, embeddable, content, teamId, content, statsEventId, language, Y, Y, highlight, 6, realTimeHighlight, 235, en, Y, Y, highlight, 6, Real-Time Highlight, 235, English
#> 2 shareable, embeddable, content, statsEventId, teamId, content, language, Y, Y, highlight, 473, 6, realTimeHighlight, en, Y, Y, highlight, 473, 6, Real-Time Highlight, English
#> 3 shareable, statsEventId, teamId, embeddable, content, content, language, Y, 462, 9, Y, highlight, realTimeHighlight, en, Y, 462, 9, Y, highlight, Real-Time Highlight, English
#> 4 shareable, embeddable, content, statsEventId, teamId, content, language, Y, Y, highlight, 451, 6, realTimeHighlight, en, Y, Y, highlight, 451, 6, Real-Time Highlight, English
#> 5   shareable, teamId, embeddable, statsEventId, content, content, language, Y, 9, Y, 58, highlight, realTimeHighlight, en, Y, 9, Y, 58, highlight, Real-Time Highlight, English
#> 6 shareable, statsEventId, teamId, embeddable, content, content, language, Y, 463, 9, Y, highlight, realTimeHighlight, en, Y, 463, 9, Y, highlight, Real-Time Highlight, English
#>   image.title image.altText image.cuts.1136x640.aspectRatio
#> 1                                                      16:9
#> 2                                                      16:9
#> 3                                                      16:9
#> 4                                                      16:9
#> 5                                                      16:9
#> 6                                                      16:9
#>   image.cuts.1136x640.width image.cuts.1136x640.height
#> 1                      1136                        640
#> 2                      1136                        640
#> 3                      1136                        640
#> 4                      1136                        640
#> 5                      1136                        640
#> 6                      1136                        640
#>                                                image.cuts.1136x640.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/1136x640/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/1136x640/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/1136x640/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/1136x640/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/1136x640/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/1136x640/cut.jpg
#>                                               image.cuts.1136x640.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/1136x640/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/1136x640/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/1136x640/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/1136x640/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/1136x640/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/1136x640/cut.jpg
#>                                               image.cuts.1136x640.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/1136x640/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/1136x640/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/1136x640/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/1136x640/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/1136x640/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/1136x640/cut.jpg
#>   image.cuts.1024x576.aspectRatio image.cuts.1024x576.width
#> 1                            16:9                      1024
#> 2                            16:9                      1024
#> 3                            16:9                      1024
#> 4                            16:9                      1024
#> 5                            16:9                      1024
#> 6                            16:9                      1024
#>   image.cuts.1024x576.height
#> 1                        576
#> 2                        576
#> 3                        576
#> 4                        576
#> 5                        576
#> 6                        576
#>                                                image.cuts.1024x576.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/1024x576/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/1024x576/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/1024x576/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/1024x576/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/1024x576/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/1024x576/cut.jpg
#>                                               image.cuts.1024x576.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/1024x576/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/1024x576/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/1024x576/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/1024x576/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/1024x576/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/1024x576/cut.jpg
#>                                               image.cuts.1024x576.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/1024x576/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/1024x576/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/1024x576/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/1024x576/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/1024x576/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/1024x576/cut.jpg
#>   image.cuts.960x540.aspectRatio image.cuts.960x540.width
#> 1                           16:9                      960
#> 2                           16:9                      960
#> 3                           16:9                      960
#> 4                           16:9                      960
#> 5                           16:9                      960
#> 6                           16:9                      960
#>   image.cuts.960x540.height
#> 1                       540
#> 2                       540
#> 3                       540
#> 4                       540
#> 5                       540
#> 6                       540
#>                                                image.cuts.960x540.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/960x540/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/960x540/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/960x540/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/960x540/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/960x540/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/960x540/cut.jpg
#>                                               image.cuts.960x540.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/960x540/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/960x540/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/960x540/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/960x540/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/960x540/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/960x540/cut.jpg
#>                                               image.cuts.960x540.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/960x540/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/960x540/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/960x540/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/960x540/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/960x540/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/960x540/cut.jpg
#>   image.cuts.768x432.aspectRatio image.cuts.768x432.width
#> 1                           16:9                      768
#> 2                           16:9                      768
#> 3                           16:9                      768
#> 4                           16:9                      768
#> 5                           16:9                      768
#> 6                           16:9                      768
#>   image.cuts.768x432.height
#> 1                       432
#> 2                       432
#> 3                       432
#> 4                       432
#> 5                       432
#> 6                       432
#>                                                image.cuts.768x432.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/768x432/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/768x432/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/768x432/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/768x432/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/768x432/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/768x432/cut.jpg
#>                                               image.cuts.768x432.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/768x432/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/768x432/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/768x432/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/768x432/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/768x432/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/768x432/cut.jpg
#>                                               image.cuts.768x432.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/768x432/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/768x432/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/768x432/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/768x432/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/768x432/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/768x432/cut.jpg
#>   image.cuts.640x360.aspectRatio image.cuts.640x360.width
#> 1                           16:9                      640
#> 2                           16:9                      640
#> 3                           16:9                      640
#> 4                           16:9                      640
#> 5                           16:9                      640
#> 6                           16:9                      640
#>   image.cuts.640x360.height
#> 1                       360
#> 2                       360
#> 3                       360
#> 4                       360
#> 5                       360
#> 6                       360
#>                                                image.cuts.640x360.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/640x360/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/640x360/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/640x360/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/640x360/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/640x360/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/640x360/cut.jpg
#>                                               image.cuts.640x360.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/640x360/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/640x360/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/640x360/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/640x360/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/640x360/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/640x360/cut.jpg
#>                                               image.cuts.640x360.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/640x360/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/640x360/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/640x360/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/640x360/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/640x360/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/640x360/cut.jpg
#>   image.cuts.568x320.aspectRatio image.cuts.568x320.width
#> 1                           16:9                      568
#> 2                           16:9                      568
#> 3                           16:9                      568
#> 4                           16:9                      568
#> 5                           16:9                      568
#> 6                           16:9                      568
#>   image.cuts.568x320.height
#> 1                       320
#> 2                       320
#> 3                       320
#> 4                       320
#> 5                       320
#> 6                       320
#>                                                image.cuts.568x320.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/568x320/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/568x320/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/568x320/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/568x320/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/568x320/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/568x320/cut.jpg
#>                                                image.cuts.568x320.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/1136x640/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/1136x640/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/1136x640/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/1136x640/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/1136x640/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/1136x640/cut.jpg
#>                                               image.cuts.568x320.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/568x320/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/568x320/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/568x320/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/568x320/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/568x320/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/568x320/cut.jpg
#>   image.cuts.372x210.aspectRatio image.cuts.372x210.width
#> 1                           16:9                      372
#> 2                           16:9                      372
#> 3                           16:9                      372
#> 4                           16:9                      372
#> 5                           16:9                      372
#> 6                           16:9                      372
#>   image.cuts.372x210.height
#> 1                       210
#> 2                       210
#> 3                       210
#> 4                       210
#> 5                       210
#> 6                       210
#>                                                image.cuts.372x210.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/372x210/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/372x210/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/372x210/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/372x210/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/372x210/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/372x210/cut.jpg
#>                                               image.cuts.372x210.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/372x210/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/372x210/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/372x210/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/372x210/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/372x210/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/372x210/cut.jpg
#>                                               image.cuts.372x210.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/372x210/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/372x210/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/372x210/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/372x210/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/372x210/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/372x210/cut.jpg
#>   image.cuts.320x180.aspectRatio image.cuts.320x180.width
#> 1                           16:9                      320
#> 2                           16:9                      320
#> 3                           16:9                      320
#> 4                           16:9                      320
#> 5                           16:9                      320
#> 6                           16:9                      320
#>   image.cuts.320x180.height
#> 1                       180
#> 2                       180
#> 3                       180
#> 4                       180
#> 5                       180
#> 6                       180
#>                                                image.cuts.320x180.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/320x180/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/320x180/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/320x180/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/320x180/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/320x180/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/320x180/cut.jpg
#>                                               image.cuts.320x180.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/640x360/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/640x360/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/640x360/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/640x360/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/640x360/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/640x360/cut.jpg
#>                                               image.cuts.320x180.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/960x540/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/960x540/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/960x540/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/960x540/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/960x540/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/960x540/cut.jpg
#>   image.cuts.248x140.aspectRatio image.cuts.248x140.width
#> 1                           16:9                      248
#> 2                           16:9                      248
#> 3                           16:9                      248
#> 4                           16:9                      248
#> 5                           16:9                      248
#> 6                           16:9                      248
#>   image.cuts.248x140.height
#> 1                       140
#> 2                       140
#> 3                       140
#> 4                       140
#> 5                       140
#> 6                       140
#>                                                image.cuts.248x140.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/248x140/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/248x140/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/248x140/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/248x140/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/248x140/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/248x140/cut.jpg
#>                                               image.cuts.248x140.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/248x140/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/248x140/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/248x140/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/248x140/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/248x140/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/248x140/cut.jpg
#>                                               image.cuts.248x140.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/248x140/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/248x140/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/248x140/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/248x140/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/248x140/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/248x140/cut.jpg
#>   image.cuts.124x70.aspectRatio image.cuts.124x70.width
#> 1                          16:9                     124
#> 2                          16:9                     124
#> 3                          16:9                     124
#> 4                          16:9                     124
#> 5                          16:9                     124
#> 6                          16:9                     124
#>   image.cuts.124x70.height
#> 1                       70
#> 2                       70
#> 3                       70
#> 4                       70
#> 5                       70
#> 6                       70
#>                                                image.cuts.124x70.src
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/124x70/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/124x70/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/124x70/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/124x70/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/124x70/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/124x70/cut.jpg
#>                                                image.cuts.124x70.at2x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/248x140/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/248x140/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/248x140/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/248x140/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/248x140/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/248x140/cut.jpg
#>                                                image.cuts.124x70.at3x
#> 1 https://cms.nhl.bamgrid.com/images/photos/327757952/372x210/cut.jpg
#> 2 https://cms.nhl.bamgrid.com/images/photos/327760402/372x210/cut.jpg
#> 3 https://cms.nhl.bamgrid.com/images/photos/327759760/372x210/cut.jpg
#> 4 https://cms.nhl.bamgrid.com/images/photos/327758654/372x210/cut.jpg
#> 5 https://cms.nhl.bamgrid.com/images/photos/327755690/372x210/cut.jpg
#> 6 https://cms.nhl.bamgrid.com/images/photos/327760182/372x210/cut.jpg
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             playbacks
#> 1 FLASH_192K_320X180, FLASH_450K_384x216, FLASH_1200K_640X360, FLASH_1800K_896x504, HTTP_CLOUD_MOBILE, HTTP_CLOUD_TABLET, HTTP_CLOUD_TABLET_60, HTTP_CLOUD_WIRED, HTTP_CLOUD_WIRED_60, HTTP_CLOUD_WIRED_WEB, 320, NA, 640, NA, null, null, NA, null, NA, NA, 180, NA, 360, NA, null, null, NA, null, NA, NA, https://hlslive-wsczoominwestus.med.nhl.com/editor/d2832146-a67c-432c-81cd-b497927e2e70.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/59490aee-157d-4b77-a509-ea76aaa3e99e.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/fa9d80cf-8c74-49d4-ad28-38aa4c5f4413.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/9ca0e463-840b-42ad-b9a9-b5a235fff7d2.mp4, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524433/MasterMobile.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524433/MasterTablet.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524433/MasterTablet60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524433/MasterWired.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524433/MasterWired60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524433/MasterWiredWeb.m3u8
#> 2 FLASH_192K_320X180, FLASH_450K_384x216, FLASH_1200K_640X360, FLASH_1800K_896x504, HTTP_CLOUD_MOBILE, HTTP_CLOUD_TABLET, HTTP_CLOUD_TABLET_60, HTTP_CLOUD_WIRED, HTTP_CLOUD_WIRED_60, HTTP_CLOUD_WIRED_WEB, 320, NA, 640, NA, null, null, NA, null, NA, NA, 180, NA, 360, NA, null, null, NA, null, NA, NA, https://hlslive-wsczoominwestus.med.nhl.com/editor/5bceb9ae-4703-4b6f-b3bf-cdfa5d3fcdf9.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/1a0c8246-6f26-49db-8afa-bc6de1ef019f.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/f5136b13-399f-44c2-8794-7d650f46c2f3.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/c9359d93-9092-4993-853f-dba7114a8c34.mp4, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525064/MasterMobile.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525064/MasterTablet.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525064/MasterTablet60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525064/MasterWired.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525064/MasterWired60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525064/MasterWiredWeb.m3u8
#> 3 FLASH_192K_320X180, FLASH_450K_384x216, FLASH_1200K_640X360, FLASH_1800K_896x504, HTTP_CLOUD_MOBILE, HTTP_CLOUD_TABLET, HTTP_CLOUD_TABLET_60, HTTP_CLOUD_WIRED, HTTP_CLOUD_WIRED_60, HTTP_CLOUD_WIRED_WEB, 320, NA, 640, NA, null, null, NA, null, NA, NA, 180, NA, 360, NA, null, null, NA, null, NA, NA, https://hlslive-wsczoominwestus.med.nhl.com/editor/faed346d-6535-4dd0-9c9b-85a2a3410eb4.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/5f60f675-0d3f-4b56-b39d-37b403119894.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/99b81ba7-5a0d-4ce1-a1ea-aa50fd9184bb.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/5a9c650e-e5b1-4183-a344-3744349a6aa6.mp4, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524883/MasterMobile.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524883/MasterTablet.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524883/MasterTablet60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524883/MasterWired.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524883/MasterWired60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524883/MasterWiredWeb.m3u8
#> 4 FLASH_192K_320X180, FLASH_450K_384x216, FLASH_1200K_640X360, FLASH_1800K_896x504, HTTP_CLOUD_MOBILE, HTTP_CLOUD_TABLET, HTTP_CLOUD_TABLET_60, HTTP_CLOUD_WIRED, HTTP_CLOUD_WIRED_60, HTTP_CLOUD_WIRED_WEB, 320, NA, 640, NA, null, null, NA, null, NA, NA, 180, NA, 360, NA, null, null, NA, null, NA, NA, https://hlslive-wsczoominwestus.med.nhl.com/editor/e4395e05-d28e-46fa-8aa7-e84fb400e594.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/19ae984c-ad0b-4e6a-9348-71a27e423309.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/ca2b9f02-3ae8-46c8-b045-b05a3df1d4ad.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/0e7214da-ddfd-460d-b6a5-64e2c6ad6a07.mp4, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524683/MasterMobile.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524683/MasterTablet.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524683/MasterTablet60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524683/MasterWired.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524683/MasterWired60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3524683/MasterWiredWeb.m3u8
#> 5 FLASH_192K_320X180, FLASH_450K_384x216, FLASH_1200K_640X360, FLASH_1800K_896x504, HTTP_CLOUD_MOBILE, HTTP_CLOUD_TABLET, HTTP_CLOUD_TABLET_60, HTTP_CLOUD_WIRED, HTTP_CLOUD_WIRED_60, HTTP_CLOUD_WIRED_WEB, 320, NA, 640, NA, null, null, NA, null, NA, NA, 180, NA, 360, NA, null, null, NA, null, NA, NA, https://hlslive-wsczoominwestus.med.nhl.com/editor/5549fb94-c970-41f5-ae9b-b2577a70fa21.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/e1da2835-2515-4d9b-b237-56bc88e25598.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/e029b4f2-42b7-46bb-a481-32b9ee2cbed8.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/fb1ec2f1-e19c-4e5e-8435-04f007c09619.mp4, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3523549/MasterMobile.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3523549/MasterTablet.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3523549/MasterTablet60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3523549/MasterWired.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3523549/MasterWired60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3523549/MasterWiredWeb.m3u8
#> 6 FLASH_192K_320X180, FLASH_450K_384x216, FLASH_1200K_640X360, FLASH_1800K_896x504, HTTP_CLOUD_MOBILE, HTTP_CLOUD_TABLET, HTTP_CLOUD_TABLET_60, HTTP_CLOUD_WIRED, HTTP_CLOUD_WIRED_60, HTTP_CLOUD_WIRED_WEB, 320, NA, 640, NA, null, null, NA, null, NA, NA, 180, NA, 360, NA, null, null, NA, null, NA, NA, https://hlslive-wsczoominwestus.med.nhl.com/editor/691fcc01-2225-40b0-9687-62a37409854e.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/706fe77f-737a-4050-891e-0801919b40ed.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/f3ef96fa-70d3-44de-b63d-a31e172014a4.mp4, https://hlslive-wsczoominwestus.med.nhl.com/editor/70c9d693-3985-42b5-b96e-d1dc14e928af.mp4, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525003/MasterMobile.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525003/MasterTablet.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525003/MasterTablet60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525003/MasterWired.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525003/MasterWired60.m3u8, https://hlslive-wsczoominwestus.med.nhl.com/publish-hls/3525003/MasterWiredWeb.m3u8
# }