NHL Player Stats

Returns player stats for a given player ID

nhl_player_stats(player_id)

Arguments

player_id
Player unique ID

Value

Returns a named list of data frames: player, stats

Examples

# \donttest{
  nhl_player_stats(player_id=8476899)
#> $player
#>   player_id   full_name                   link first_name last_name
#> 1   8476899 Matt Murray /api/v1/people/8476899       Matt    Murray
#>   primary_number birth_date current_age  birth_city birth_state_province
#> 1             30 1994-05-25          27 Thunder Bay                   ON
#>   birth_country nationality height weight active alternate_captain captain
#> 1           CAN         CAN  6' 4"    202   TRUE             FALSE   FALSE
#>   rookie shoots_catches roster_status
#> 1  FALSE              L             Y
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           stats
#> 1 20072008, 20082009, 20092010, 20102011, 20112012, 20112012, 20122013, 20132014, 20132014, 20142015, 20142015, 20152016, 20152016, 20162017, 20162017, 20172018, 20182019, 20182019, 20192020, 20202021, 20212022, 129295, 129295, 1, 1, 1, 2, 1, 1, 2, 1, 1077, 1, 1077, 1, 176776, 1, 1, 17155, 1, 1, 1, 0.87, 0, NA, NA, NA, 0.9095, 0.8942, 0.9212, 0.92, 0.941, 0.941, 0.9296, 0.931, 0.9234, 0.886, 0.9065, 0.9191, 0.926, 0.899, 0.893, 0.897, 5.92, 2.84, 1.71, 3.79, 4.08, 2.72, 3.67, 2.57, 2, 1.58, 1.58, 2.0038, 2.1, 2.408, 3.16, 2.9199, 2.6873, 2.01, 2.8693, 3.3756, 3.1036, 32, 29, 40, 28, 36, 7, 53, 49, 1, 40, 40, 13, 31, 49, 2, 49, 50, 7, 38, 27, 5, NA, NA, 1975:00, 1377:00, 1912:00, 420:00, 2910:00, 2984:00, 60:00, 2321:00, NA, 748:35, NA, 2765:48, NA, 2732:58, 2880:14, NA, 2237:30, 1404:11, 289:59, NA, NA, 6, 1, 0, 0, 2, 6, 0, 12, NA, 1, NA, 4, NA, 1, 4, NA, 1, 2, 0, NA, NA, 0, 3, 1, 0, 4, 1, 0, 3, NA, NA, NA, NA, NA, 0, 0, NA, 0, 0, 0, NA, NA, 32, 8, 13, 4, 26, 32, 0, 25, NA, 9, NA, 32, NA, 27, 29, NA, 20, 10, 0, NA, NA, 5, 11, 19, 3, 19, 11, 1, 10, NA, 2, NA, 10, NA, 16, 14, NA, 11, 13, 4, NA, NA, 75, 87, 130, 19, 178, 128, 2, 61, NA, 25, NA, 111, NA, 133, 129, NA, 107, 79, 15, NA, NA, NA, NA, NA, 191, 1504, 1497, 23, 968, NA, 330, NA, 1339, NA, 1290, 1465, NA, 948, 659, 131, NA, NA, NA, NA, NA, 210, 1682, 1625, 25, 1029, NA, 355, NA, 1450, NA, 1423, 1594, NA, 1055, 738, 146, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1, NA, 4, NA, 3, 6, NA, 5, 1, 0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 54, NA, 182, NA, 192, 166, NA, 111, 102, 22, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 10, NA, 52, NA, 38, 41, NA, 34, 17, 4, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 266, NA, 1105, NA, 1060, 1258, NA, 803, 540, 105, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 11, NA, 58, NA, 39, 48, NA, 37, 21, 4, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 285, NA, 1185, NA, 1163, 1352, NA, 891, 599, 115, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 59, NA, 207, NA, 221, 194, NA, 127, 118, 27, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 13, NA, 47, NA, 45, 50, NA, 38, 25, 5, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 91.5254, NA, 87.9227, NA, 86.8778, 85.567, NA, 87.4016, 86.4407, 81.4815, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 90.9091, NA, 89.6552, NA, 97.4359, 85.4167, NA, 91.8919, 80.9524, 100, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 93.3333, NA, 93.2489, NA, 91.1436, 93.0473, NA, 90.1235, 90.1503, 91.3043, Thunder Bay Kings Bantam AAA, Thunder Bay Kings Bantam AAA, Thunder Bay, Sault Ste. Marie, Sault Ste. Marie, Canada, Sault Ste. Marie, Sault Ste. Marie, Wilkes-Barre, Wilkes-Barre, Wilkes-Barre/Scranton Penguins, Pittsburgh Penguins, Wilkes-Barre/Scranton Penguins, Pittsburgh Penguins, Team North America, Pittsburgh Penguins, Pittsburgh Penguins, Canada, Pittsburgh Penguins, Ottawa Senators, Ottawa Senators, /api/v1/teams/null, /api/v1/teams/null, /api/v1/teams/null, /api/v1/teams/2762, /api/v1/teams/2762, /api/v1/teams/60, /api/v1/teams/2762, /api/v1/teams/2762, /api/v1/teams/6054, /api/v1/teams/6054, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/5, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/9, /api/v1/teams/9, NA, NA, NA, 2762, 2762, 60, 2762, 2762, 6054, 6054, NA, 5, NA, 5, NA, 5, 5, NA, 5, 9, 9, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, PIT, NA, PIT, NA, PIT, PIT, NA, PIT, OTT, OTT, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Penguins, NA, Penguins, NA, Penguins, Penguins, NA, Penguins, Senators, Senators, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Pittsburgh, NA, Pittsburgh, NA, Pittsburgh, Pittsburgh, NA, Pittsburgh, Ottawa, Ottawa, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1967, NA, 1967, NA, 1967, 1967, NA, 1967, 1992, 1992, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Pittsburgh, NA, Pittsburgh, NA, Pittsburgh, Pittsburgh, NA, Pittsburgh, Ottawa, Ottawa, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, http://pittsburghpenguins.com/, NA, http://pittsburghpenguins.com/, NA, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, NA, http://pittsburghpenguins.com/, http://www.ottawasenators.com/, http://www.ottawasenators.com/, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 17, NA, 17, NA, 17, 17, NA, 17, 30, 30, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, TRUE, NA, TRUE, NA, TRUE, TRUE, NA, TRUE, TRUE, TRUE, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 5034, NA, 5034, NA, 5034, 5034, NA, 5034, 5031, 5031, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, PPG Paints Arena, NA, PPG Paints Arena, NA, PPG Paints Arena, PPG Paints Arena, NA, PPG Paints Arena, Canadian Tire Centre, Canadian Tire Centre, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, /api/v1/venues/5034, NA, /api/v1/venues/5034, NA, /api/v1/venues/5034, /api/v1/venues/5034, NA, /api/v1/venues/5034, /api/v1/venues/5031, /api/v1/venues/5031, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Pittsburgh, NA, Pittsburgh, NA, Pittsburgh, Pittsburgh, NA, Pittsburgh, Ottawa, Ottawa, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, America/New_York, NA, America/New_York, NA, America/New_York, America/New_York, NA, America/New_York, America/New_York, America/New_York, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, -5, NA, -5, NA, -5, -5, NA, -5, -5, -5, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, EST, NA, EST, NA, EST, EST, NA, EST, EST, EST, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 18, NA, 18, NA, 18, 18, NA, 18, 28, 17, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Metropolitan, NA, Metropolitan, NA, Metropolitan, Metropolitan, NA, Metropolitan, Scotia North, Atlantic, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Metro, NA, Metro, NA, Metro, Metro, NA, Metro, NA, ATL, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, /api/v1/divisions/18, NA, /api/v1/divisions/18, NA, /api/v1/divisions/18, /api/v1/divisions/18, NA, /api/v1/divisions/18, /api/v1/divisions/28, /api/v1/divisions/17, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, M, NA, M, NA, M, M, NA, M, NA, A, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 6, NA, 6, NA, 6, 6, NA, 6, 6, 6, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Eastern, NA, Eastern, NA, Eastern, Eastern, NA, Eastern, Eastern, Eastern, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, /api/v1/conferences/6, NA, /api/v1/conferences/6, NA, /api/v1/conferences/6, /api/v1/conferences/6, NA, /api/v1/conferences/6, /api/v1/conferences/6, /api/v1/conferences/6, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 17, NA, 17, NA, 17, 17, NA, 17, 30, 30, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Penguins, NA, Penguins, NA, Penguins, Penguins, NA, Penguins, Senators, Senators, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, /api/v1/franchises/17, NA, /api/v1/franchises/17, NA, /api/v1/franchises/17, /api/v1/franchises/17, NA, /api/v1/franchises/17, /api/v1/franchises/30, /api/v1/franchises/30, Bantam, Bantam, Minor-ON, OHL, OHL, WJ18-A, OHL, OHL, AHL, AHL, AHL, National Hockey League, AHL, National Hockey League, WCup, National Hockey League, National Hockey League, WC, National Hockey League, National Hockey League, National Hockey League, /api/v1/league/null, /api/v1/league/null, /api/v1/league/null, /api/v1/league/141, /api/v1/league/141, /api/v1/league/147, /api/v1/league/141, /api/v1/league/141, /api/v1/league/153, /api/v1/league/153, /api/v1/league/null, /api/v1/league/133, /api/v1/league/null, /api/v1/league/133, /api/v1/league/null, /api/v1/league/133, /api/v1/league/133, /api/v1/league/null, /api/v1/league/133, /api/v1/league/133, /api/v1/league/133, NA, NA, NA, 141, 141, 147, 141, 141, 153, 153, NA, 133, NA, 133, NA, 133, 133, NA, 133, 133, 133, 20122013, 20132014, 20132014, 20142015, 20142015, 20152016, 20162017, 20172018, 20182019, 20192020, 1, 1, 2, 1, 1077, 1, 1, 1, 1, 1, 381:00, 547:00, 20:00, 456:00, NA, 1267:16, 668:55, 715:35, 238:44, 192:07, 1, 1, 0, 1, NA, 1, 3, 2, 0, 0, 2, 4, 0, 4, NA, 15, 7, 6, 0, 1, 4, 3, 0, 4, NA, 6, 3, 6, 4, 2, 171, 257, 8, 217, NA, 531, 284, 285, 115, 85, 0.9096, 0.9146, 1, 0.923, 0.923, 0.9235, 0.9373, 0.9076, 0.9055, 0.914, 2.67, 2.63, 0, 2.37, 2.37, 2.0832, 1.7042, 2.4316, 3.0159, 2.4985, 6, 9, 1, 8, 8, 21, 11, 12, 4, 3, 188, 281, 8, 235, NA, 575, 303, 314, 127, 93, 17, 24, 0, 18, NA, 44, 19, 29, 12, 8, NA, 2, 0, 0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 68, 36, 33, 21, 8, NA, NA, NA, NA, NA, 5, 13, 19, 7, 4, NA, NA, NA, NA, NA, 458, 235, 233, 87, 73, NA, NA, NA, NA, NA, 6, 13, 20, 7, 4, NA, NA, NA, NA, NA, 492, 249, 255, 97, 81, NA, NA, NA, NA, NA, 77, 41, 39, 23, 8, NA, NA, NA, NA, NA, 21, 10, 12, 4, 3, NA, NA, NA, NA, NA, 88.3117, 87.8049, 84.6154, 91.3043, 100, NA, NA, NA, NA, NA, 83.3333, 100, 95, 100, 100, NA, NA, NA, NA, NA, 93.0894, 94.3775, 91.3725, 89.6907, 90.1235, NA, NA, NA, NA, NA, NA, NA, NA, NA, 0, 2762, 2762, 6054, 6054, NA, 5, 5, 5, 5, 5, Sault Ste. Marie, Sault Ste. Marie, Wilkes-Barre, Wilkes-Barre, Wilkes-Barre/Scranton Penguins, Pittsburgh Penguins, Pittsburgh Penguins, Pittsburgh Penguins, Pittsburgh Penguins, Pittsburgh Penguins, /api/v1/teams/2762, /api/v1/teams/2762, /api/v1/teams/6054, /api/v1/teams/6054, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/5, /api/v1/teams/5, /api/v1/teams/5, /api/v1/teams/5, NA, NA, NA, NA, NA, PIT, PIT, PIT, PIT, PIT, NA, NA, NA, NA, NA, Penguins, Penguins, Penguins, Penguins, Penguins, NA, NA, NA, NA, NA, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, NA, NA, NA, NA, NA, 1967, 1967, 1967, 1967, 1967, NA, NA, NA, NA, NA, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, NA, NA, NA, NA, NA, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, NA, NA, NA, NA, NA, 17, 17, 17, 17, 17, NA, NA, NA, NA, NA, TRUE, TRUE, TRUE, TRUE, TRUE, NA, NA, NA, NA, NA, 5034, 5034, 5034, 5034, 5034, NA, NA, NA, NA, NA, PPG Paints Arena, PPG Paints Arena, PPG Paints Arena, PPG Paints Arena, PPG Paints Arena, NA, NA, NA, NA, NA, /api/v1/venues/5034, /api/v1/venues/5034, /api/v1/venues/5034, /api/v1/venues/5034, /api/v1/venues/5034, NA, NA, NA, NA, NA, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, NA, NA, NA, NA, NA, America/New_York, America/New_York, America/New_York, America/New_York, America/New_York, NA, NA, NA, NA, NA, -5, -5, -5, -5, -5, NA, NA, NA, NA, NA, EST, EST, EST, EST, EST, NA, NA, NA, NA, NA, 18, 18, 18, 18, 18, NA, NA, NA, NA, NA, Metropolitan, Metropolitan, Metropolitan, Metropolitan, Metropolitan, NA, NA, NA, NA, NA, Metro, Metro, Metro, Metro, Metro, NA, NA, NA, NA, NA, /api/v1/divisions/18, /api/v1/divisions/18, /api/v1/divisions/18, /api/v1/divisions/18, /api/v1/divisions/18, NA, NA, NA, NA, NA, M, M, M, M, M, NA, NA, NA, NA, NA, 6, 6, 6, 6, 6, NA, NA, NA, NA, NA, Eastern, Eastern, Eastern, Eastern, Eastern, NA, NA, NA, NA, NA, /api/v1/conferences/6, /api/v1/conferences/6, /api/v1/conferences/6, /api/v1/conferences/6, /api/v1/conferences/6, NA, NA, NA, NA, NA, 17, 17, 17, 17, 17, NA, NA, NA, NA, NA, Penguins, Penguins, Penguins, Penguins, Penguins, NA, NA, NA, NA, NA, /api/v1/franchises/17, /api/v1/franchises/17, /api/v1/franchises/17, /api/v1/franchises/17, /api/v1/franchises/17, 141, 141, 153, 153, NA, 133, 133, 133, 133, 133, OHL, OHL, AHL, AHL, AHL, National Hockey League, National Hockey League, National Hockey League, National Hockey League, National Hockey League, /api/v1/league/141, /api/v1/league/141, /api/v1/league/153, /api/v1/league/153, /api/v1/league/null, /api/v1/league/133, /api/v1/league/133, /api/v1/league/133, /api/v1/league/133, /api/v1/league/133, 13059:15, 20, 13, 127, 70, 6162, 829, 196, 5137, 218, 5590, 953, 0.9114, 2.7521, 231, 223, 6761, 599, 56:32, 86.9885, 89.9083, 91.8962, yearByYear, yearByYearPlayoffs, careerRegularSeason
#>   current_team_id current_team_name current_team_link primary_position_code
#> 1               9   Ottawa Senators   /api/v1/teams/9                     G
#>   primary_position_name primary_position_type primary_position_abbreviation
#> 1                Goalie                Goalie                             G
#> 
#> $stats
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         splits
#> 1 20072008, 20082009, 20092010, 20102011, 20112012, 20112012, 20122013, 20132014, 20132014, 20142015, 20142015, 20152016, 20152016, 20162017, 20162017, 20172018, 20182019, 20182019, 20192020, 20202021, 20212022, 129295, 129295, 1, 1, 1, 2, 1, 1, 2, 1, 1077, 1, 1077, 1, 176776, 1, 1, 17155, 1, 1, 1, 0.87, 0, NA, NA, NA, 0.9095, 0.8942, 0.9212, 0.92, 0.941, 0.941, 0.9296, 0.931, 0.9234, 0.886, 0.9065, 0.9191, 0.926, 0.899, 0.893, 0.897, 5.92, 2.84, 1.71, 3.79, 4.08, 2.72, 3.67, 2.57, 2, 1.58, 1.58, 2.0038, 2.1, 2.408, 3.16, 2.9199, 2.6873, 2.01, 2.8693, 3.3756, 3.1036, 32, 29, 40, 28, 36, 7, 53, 49, 1, 40, 40, 13, 31, 49, 2, 49, 50, 7, 38, 27, 5, NA, NA, 1975:00, 1377:00, 1912:00, 420:00, 2910:00, 2984:00, 60:00, 2321:00, NA, 748:35, NA, 2765:48, NA, 2732:58, 2880:14, NA, 2237:30, 1404:11, 289:59, NA, NA, 6, 1, 0, 0, 2, 6, 0, 12, NA, 1, NA, 4, NA, 1, 4, NA, 1, 2, 0, NA, NA, 0, 3, 1, 0, 4, 1, 0, 3, NA, NA, NA, NA, NA, 0, 0, NA, 0, 0, 0, NA, NA, 32, 8, 13, 4, 26, 32, 0, 25, NA, 9, NA, 32, NA, 27, 29, NA, 20, 10, 0, NA, NA, 5, 11, 19, 3, 19, 11, 1, 10, NA, 2, NA, 10, NA, 16, 14, NA, 11, 13, 4, NA, NA, 75, 87, 130, 19, 178, 128, 2, 61, NA, 25, NA, 111, NA, 133, 129, NA, 107, 79, 15, NA, NA, NA, NA, NA, 191, 1504, 1497, 23, 968, NA, 330, NA, 1339, NA, 1290, 1465, NA, 948, 659, 131, NA, NA, NA, NA, NA, 210, 1682, 1625, 25, 1029, NA, 355, NA, 1450, NA, 1423, 1594, NA, 1055, 738, 146, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1, NA, 4, NA, 3, 6, NA, 5, 1, 0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 54, NA, 182, NA, 192, 166, NA, 111, 102, 22, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 10, NA, 52, NA, 38, 41, NA, 34, 17, 4, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 266, NA, 1105, NA, 1060, 1258, NA, 803, 540, 105, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 11, NA, 58, NA, 39, 48, NA, 37, 21, 4, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 285, NA, 1185, NA, 1163, 1352, NA, 891, 599, 115, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 59, NA, 207, NA, 221, 194, NA, 127, 118, 27, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 13, NA, 47, NA, 45, 50, NA, 38, 25, 5, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 91.5254, NA, 87.9227, NA, 86.8778, 85.567, NA, 87.4016, 86.4407, 81.4815, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 90.9091, NA, 89.6552, NA, 97.4359, 85.4167, NA, 91.8919, 80.9524, 100, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 93.3333, NA, 93.2489, NA, 91.1436, 93.0473, NA, 90.1235, 90.1503, 91.3043, Thunder Bay Kings Bantam AAA, Thunder Bay Kings Bantam AAA, Thunder Bay, Sault Ste. Marie, Sault Ste. Marie, Canada, Sault Ste. Marie, Sault Ste. Marie, Wilkes-Barre, Wilkes-Barre, Wilkes-Barre/Scranton Penguins, Pittsburgh Penguins, Wilkes-Barre/Scranton Penguins, Pittsburgh Penguins, Team North America, Pittsburgh Penguins, Pittsburgh Penguins, Canada, Pittsburgh Penguins, Ottawa Senators, Ottawa Senators, /api/v1/teams/null, /api/v1/teams/null, /api/v1/teams/null, /api/v1/teams/2762, /api/v1/teams/2762, /api/v1/teams/60, /api/v1/teams/2762, /api/v1/teams/2762, /api/v1/teams/6054, /api/v1/teams/6054, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/5, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/9, /api/v1/teams/9, NA, NA, NA, 2762, 2762, 60, 2762, 2762, 6054, 6054, NA, 5, NA, 5, NA, 5, 5, NA, 5, 9, 9, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, PIT, NA, PIT, NA, PIT, PIT, NA, PIT, OTT, OTT, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Penguins, NA, Penguins, NA, Penguins, Penguins, NA, Penguins, Senators, Senators, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Pittsburgh, NA, Pittsburgh, NA, Pittsburgh, Pittsburgh, NA, Pittsburgh, Ottawa, Ottawa, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 1967, NA, 1967, NA, 1967, 1967, NA, 1967, 1992, 1992, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Pittsburgh, NA, Pittsburgh, NA, Pittsburgh, Pittsburgh, NA, Pittsburgh, Ottawa, Ottawa, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, http://pittsburghpenguins.com/, NA, http://pittsburghpenguins.com/, NA, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, NA, http://pittsburghpenguins.com/, http://www.ottawasenators.com/, http://www.ottawasenators.com/, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 17, NA, 17, NA, 17, 17, NA, 17, 30, 30, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, TRUE, NA, TRUE, NA, TRUE, TRUE, NA, TRUE, TRUE, TRUE, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 5034, NA, 5034, NA, 5034, 5034, NA, 5034, 5031, 5031, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, PPG Paints Arena, NA, PPG Paints Arena, NA, PPG Paints Arena, PPG Paints Arena, NA, PPG Paints Arena, Canadian Tire Centre, Canadian Tire Centre, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, /api/v1/venues/5034, NA, /api/v1/venues/5034, NA, /api/v1/venues/5034, /api/v1/venues/5034, NA, /api/v1/venues/5034, /api/v1/venues/5031, /api/v1/venues/5031, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Pittsburgh, NA, Pittsburgh, NA, Pittsburgh, Pittsburgh, NA, Pittsburgh, Ottawa, Ottawa, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, America/New_York, NA, America/New_York, NA, America/New_York, America/New_York, NA, America/New_York, America/New_York, America/New_York, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, -5, NA, -5, NA, -5, -5, NA, -5, -5, -5, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, EST, NA, EST, NA, EST, EST, NA, EST, EST, EST, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 18, NA, 18, NA, 18, 18, NA, 18, 28, 17, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Metropolitan, NA, Metropolitan, NA, Metropolitan, Metropolitan, NA, Metropolitan, Scotia North, Atlantic, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Metro, NA, Metro, NA, Metro, Metro, NA, Metro, NA, ATL, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, /api/v1/divisions/18, NA, /api/v1/divisions/18, NA, /api/v1/divisions/18, /api/v1/divisions/18, NA, /api/v1/divisions/18, /api/v1/divisions/28, /api/v1/divisions/17, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, M, NA, M, NA, M, M, NA, M, NA, A, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 6, NA, 6, NA, 6, 6, NA, 6, 6, 6, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Eastern, NA, Eastern, NA, Eastern, Eastern, NA, Eastern, Eastern, Eastern, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, /api/v1/conferences/6, NA, /api/v1/conferences/6, NA, /api/v1/conferences/6, /api/v1/conferences/6, NA, /api/v1/conferences/6, /api/v1/conferences/6, /api/v1/conferences/6, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 17, NA, 17, NA, 17, 17, NA, 17, 30, 30, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, Penguins, NA, Penguins, NA, Penguins, Penguins, NA, Penguins, Senators, Senators, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, /api/v1/franchises/17, NA, /api/v1/franchises/17, NA, /api/v1/franchises/17, /api/v1/franchises/17, NA, /api/v1/franchises/17, /api/v1/franchises/30, /api/v1/franchises/30, Bantam, Bantam, Minor-ON, OHL, OHL, WJ18-A, OHL, OHL, AHL, AHL, AHL, National Hockey League, AHL, National Hockey League, WCup, National Hockey League, National Hockey League, WC, National Hockey League, National Hockey League, National Hockey League, /api/v1/league/null, /api/v1/league/null, /api/v1/league/null, /api/v1/league/141, /api/v1/league/141, /api/v1/league/147, /api/v1/league/141, /api/v1/league/141, /api/v1/league/153, /api/v1/league/153, /api/v1/league/null, /api/v1/league/133, /api/v1/league/null, /api/v1/league/133, /api/v1/league/null, /api/v1/league/133, /api/v1/league/133, /api/v1/league/null, /api/v1/league/133, /api/v1/league/133, /api/v1/league/133, NA, NA, NA, 141, 141, 147, 141, 141, 153, 153, NA, 133, NA, 133, NA, 133, 133, NA, 133, 133, 133
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           20122013, 20132014, 20132014, 20142015, 20142015, 20152016, 20162017, 20172018, 20182019, 20192020, 1, 1, 2, 1, 1077, 1, 1, 1, 1, 1, 381:00, 547:00, 20:00, 456:00, NA, 1267:16, 668:55, 715:35, 238:44, 192:07, 1, 1, 0, 1, NA, 1, 3, 2, 0, 0, 2, 4, 0, 4, NA, 15, 7, 6, 0, 1, 4, 3, 0, 4, NA, 6, 3, 6, 4, 2, 171, 257, 8, 217, NA, 531, 284, 285, 115, 85, 0.9096, 0.9146, 1, 0.923, 0.923, 0.9235, 0.9373, 0.9076, 0.9055, 0.914, 2.67, 2.63, 0, 2.37, 2.37, 2.0832, 1.7042, 2.4316, 3.0159, 2.4985, 6, 9, 1, 8, 8, 21, 11, 12, 4, 3, 188, 281, 8, 235, NA, 575, 303, 314, 127, 93, 17, 24, 0, 18, NA, 44, 19, 29, 12, 8, NA, 2, 0, 0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 68, 36, 33, 21, 8, NA, NA, NA, NA, NA, 5, 13, 19, 7, 4, NA, NA, NA, NA, NA, 458, 235, 233, 87, 73, NA, NA, NA, NA, NA, 6, 13, 20, 7, 4, NA, NA, NA, NA, NA, 492, 249, 255, 97, 81, NA, NA, NA, NA, NA, 77, 41, 39, 23, 8, NA, NA, NA, NA, NA, 21, 10, 12, 4, 3, NA, NA, NA, NA, NA, 88.3117, 87.8049, 84.6154, 91.3043, 100, NA, NA, NA, NA, NA, 83.3333, 100, 95, 100, 100, NA, NA, NA, NA, NA, 93.0894, 94.3775, 91.3725, 89.6907, 90.1235, NA, NA, NA, NA, NA, NA, NA, NA, NA, 0, 2762, 2762, 6054, 6054, NA, 5, 5, 5, 5, 5, Sault Ste. Marie, Sault Ste. Marie, Wilkes-Barre, Wilkes-Barre, Wilkes-Barre/Scranton Penguins, Pittsburgh Penguins, Pittsburgh Penguins, Pittsburgh Penguins, Pittsburgh Penguins, Pittsburgh Penguins, /api/v1/teams/2762, /api/v1/teams/2762, /api/v1/teams/6054, /api/v1/teams/6054, /api/v1/teams/null, /api/v1/teams/5, /api/v1/teams/5, /api/v1/teams/5, /api/v1/teams/5, /api/v1/teams/5, NA, NA, NA, NA, NA, PIT, PIT, PIT, PIT, PIT, NA, NA, NA, NA, NA, Penguins, Penguins, Penguins, Penguins, Penguins, NA, NA, NA, NA, NA, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, NA, NA, NA, NA, NA, 1967, 1967, 1967, 1967, 1967, NA, NA, NA, NA, NA, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, NA, NA, NA, NA, NA, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, http://pittsburghpenguins.com/, NA, NA, NA, NA, NA, 17, 17, 17, 17, 17, NA, NA, NA, NA, NA, TRUE, TRUE, TRUE, TRUE, TRUE, NA, NA, NA, NA, NA, 5034, 5034, 5034, 5034, 5034, NA, NA, NA, NA, NA, PPG Paints Arena, PPG Paints Arena, PPG Paints Arena, PPG Paints Arena, PPG Paints Arena, NA, NA, NA, NA, NA, /api/v1/venues/5034, /api/v1/venues/5034, /api/v1/venues/5034, /api/v1/venues/5034, /api/v1/venues/5034, NA, NA, NA, NA, NA, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, Pittsburgh, NA, NA, NA, NA, NA, America/New_York, America/New_York, America/New_York, America/New_York, America/New_York, NA, NA, NA, NA, NA, -5, -5, -5, -5, -5, NA, NA, NA, NA, NA, EST, EST, EST, EST, EST, NA, NA, NA, NA, NA, 18, 18, 18, 18, 18, NA, NA, NA, NA, NA, Metropolitan, Metropolitan, Metropolitan, Metropolitan, Metropolitan, NA, NA, NA, NA, NA, Metro, Metro, Metro, Metro, Metro, NA, NA, NA, NA, NA, /api/v1/divisions/18, /api/v1/divisions/18, /api/v1/divisions/18, /api/v1/divisions/18, /api/v1/divisions/18, NA, NA, NA, NA, NA, M, M, M, M, M, NA, NA, NA, NA, NA, 6, 6, 6, 6, 6, NA, NA, NA, NA, NA, Eastern, Eastern, Eastern, Eastern, Eastern, NA, NA, NA, NA, NA, /api/v1/conferences/6, /api/v1/conferences/6, /api/v1/conferences/6, /api/v1/conferences/6, /api/v1/conferences/6, NA, NA, NA, NA, NA, 17, 17, 17, 17, 17, NA, NA, NA, NA, NA, Penguins, Penguins, Penguins, Penguins, Penguins, NA, NA, NA, NA, NA, /api/v1/franchises/17, /api/v1/franchises/17, /api/v1/franchises/17, /api/v1/franchises/17, /api/v1/franchises/17, 141, 141, 153, 153, NA, 133, 133, 133, 133, 133, OHL, OHL, AHL, AHL, AHL, National Hockey League, National Hockey League, National Hockey League, National Hockey League, National Hockey League, /api/v1/league/141, /api/v1/league/141, /api/v1/league/153, /api/v1/league/153, /api/v1/league/null, /api/v1/league/133, /api/v1/league/133, /api/v1/league/133, /api/v1/league/133, /api/v1/league/133
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       13059:15, 20, 13, 127, 70, 6162, 829, 196, 5137, 218, 5590, 953, 0.9114, 2.7521, 231, 223, 6761, 599, 56:32, 86.9885, 89.9083, 91.8962
#>      type.displayName
#> 1          yearByYear
#> 2  yearByYearPlayoffs
#> 3 careerRegularSeason
#> 
# }